Refactored controller structure for improved design
